+++ source/blender/blenkernel/intern/action_mirror.cc
@@ -169,9 +169,9 @@ static void action_flip_pchan(Object *ob_arm, const bPoseChannel *pchan, FCurveP
   char path_xform[256];
   char pchan_name_esc[sizeof(bActionChannel::name) * 2];
   BLI_str_escape(pchan_name_esc, pchan->name, sizeof(pchan_name_esc));
-  const int path_xform_prefix_len = SNPRINTF(path_xform, "pose.bones[\"%s\"]", pchan_name_esc);
+  const size_t path_xform_prefix_len = SNPRINTF(path_xform, "pose.bones[\"%s\"]", pchan_name_esc);
   char *path_xform_suffix = path_xform + path_xform_prefix_len;
-  const int path_xform_suffix_maxncpy = sizeof(path_xform) - path_xform_prefix_len;
+  const size_t path_xform_suffix_maxncpy = sizeof(path_xform) - path_xform_prefix_len;
 
   /* Lookup and assign all available #FCurve channels,
    * unavailable channels are left nullptr. */
@@ -387,7 +387,7 @@ static void action_flip_pchan(Object *ob_arm, const bPoseChannel *pchan, FCurveP
 static void action_flip_pchan_rna_paths(bAction *act)
 {
   const char *path_pose_prefix = "pose.bones[\"";
-  const int path_pose_prefix_len = strlen(path_pose_prefix);
+  const size_t path_pose_prefix_len = strlen(path_pose_prefix);
 
   /* Tag curves that have renamed f-curves. */
   LISTBASE_FOREACH (bActionGroup *, agrp, &act->groups) {
