+++ source/blender/editors/sculpt_paint/paint_image_2d.cc
@@ -1889,7 +1889,7 @@ void paint_2d_bucket_fill(const bContext *C,
       copy_v4_v4(pixel_color, ibuf->float_buffer.data + 4 * coordinate);
     }
     else {
-      int pixel_color_b = *ibuf->byte_buffer.data + 4 * coordinate;
+      size_t pixel_color_b = *ibuf->byte_buffer.data + 4 * coordinate;
       rgba_uchar_to_float(pixel_color, (uchar *)&pixel_color_b);
       straight_to_premul_v4(pixel_color);
     }
